+++ meridian/model/model_test.py
@@ -1412,6 +1412,7 @@ class ModelTest(
           alpha=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
           ec=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
           slope=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
+          decay_functions=meridian.adstock_decay_spec.media,
       )
 
   def test_adstock_hill_media_n_times_output(self):
@@ -1430,6 +1431,7 @@ class ModelTest(
           alpha=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
           ec=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
           slope=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
+          decay_functions=meridian.adstock_decay_spec.media,
           n_times_output=8,
       )
 
@@ -1486,6 +1488,7 @@ class ModelTest(
         alpha=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
         ec=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
         slope=np.ones(shape=(self._N_MEDIA_CHANNELS,)),
+        decay_functions=meridian.adstock_decay_spec.media,
     )
 
     mock_hill.assert_called_once()
@@ -1510,6 +1513,7 @@ class ModelTest(
           alpha=np.ones(shape=(self._N_RF_CHANNELS,)),
           ec=np.ones(shape=(self._N_RF_CHANNELS,)),
           slope=np.ones(shape=(self._N_RF_CHANNELS,)),
+          decay_functions=meridian.adstock_decay_spec.rf,
       )
 
   def test_adstock_hill_rf_n_times_output(self):
@@ -1529,6 +1533,7 @@ class ModelTest(
           alpha=np.ones(shape=(self._N_RF_CHANNELS,)),
           ec=np.ones(shape=(self._N_RF_CHANNELS,)),
           slope=np.ones(shape=(self._N_RF_CHANNELS,)),
+          decay_functions=meridian.adstock_decay_spec.rf,
           n_times_output=8,
       )
 
@@ -1571,6 +1576,7 @@ class ModelTest(
         alpha=np.ones(shape=(self._N_RF_CHANNELS,)),
         ec=np.ones(shape=(self._N_RF_CHANNELS,)),
         slope=np.ones(shape=(self._N_RF_CHANNELS,)),
+        decay_functions=meridian.adstock_decay_spec.rf,
     )
 
     expected_called_names = ["mock_hill", "mock_adstock"]
@@ -2202,6 +2208,7 @@ class NonPaidModelTest(
         alpha=par[constants.ALPHA_M],
         ec=par[constants.EC_M],
         slope=par[constants.SLOPE_M],
+        decay_functions=meridian.adstock_decay_spec.media,
     )[0, :, :, :]
     transformed_reach = meridian.adstock_hill_rf(
         reach=meridian.rf_tensors.reach_scaled,
@@ -2209,12 +2216,14 @@ class NonPaidModelTest(
         alpha=par[constants.ALPHA_RF],
         ec=par[constants.EC_RF],
         slope=par[constants.SLOPE_RF],
+        decay_functions=meridian.adstock_decay_spec.rf,
     )[0, :, :, :]
     transformed_organic_media = meridian.adstock_hill_media(
         media=meridian.organic_media_tensors.organic_media_scaled,
         alpha=par[constants.ALPHA_OM],
         ec=par[constants.EC_OM],
         slope=par[constants.SLOPE_OM],
+        decay_functions=meridian.adstock_decay_spec.organic_media,
     )[0, :, :, :]
     transformed_organic_reach = meridian.adstock_hill_rf(
         reach=meridian.organic_rf_tensors.organic_reach_scaled,
@@ -2222,6 +2231,7 @@ class NonPaidModelTest(
         alpha=par[constants.ALPHA_ORF],
         ec=par[constants.EC_ORF],
         slope=par[constants.SLOPE_ORF],
+        decay_functions=meridian.adstock_decay_spec.organic_rf,
     )[0, :, :, :]
     combined_transformed_media = backend.concatenate(
         [
@@ -2707,6 +2717,228 @@ class NonPaidModelTest(
     )
     actual_baseline = meridian.compute_non_media_treatments_baseline()
     test_utils.assert_allclose(expected_baseline, actual_baseline)
+
+
+class AdstockDecaySpecFromChannelMappingTest(parameterized.TestCase):
+  @parameterized.product(**data_test_utils.ADSTOCK_DECAY_SPEC_CASES)
+  def test_from_channel(
+      self,
+      media,
+      rf,
+      organic_media,
+      organic_rf,
+  ):
+    """Test if adstock decay functions are explicitly passed for all channels."""
+
+    if not (media or rf):
+      self.skipTest("Invalid test case: Meridian requires paid media.")
+
+    inp_data = data_test_utils.sample_input_data_revenue(
+        n_media_channels=len(media),
+        n_rf_channels=len(rf),
+        n_organic_media_channels=len(organic_media),
+        n_organic_rf_channels=len(organic_rf),
+    )
+
+    decay_spec = media | rf | organic_media | organic_rf
+    model_spec = spec.ModelSpec(adstock_decay_spec=decay_spec)
+    mmm = model.Meridian(input_data=inp_data, model_spec=model_spec)
+
+    expected_media = list(media.values()) or constants.GEOMETRIC_DECAY
+    expected_rf = list(rf.values()) or constants.GEOMETRIC_DECAY
+    expected_organic_media = (
+        list(organic_media.values()) or constants.GEOMETRIC_DECAY
+    )
+    expected_organic_rf = list(organic_rf.values()) or constants.GEOMETRIC_DECAY
+
+    self.assertSequenceEqual(mmm.adstock_decay_spec.media, expected_media)
+    self.assertSequenceEqual(mmm.adstock_decay_spec.rf, expected_rf)
+    self.assertSequenceEqual(
+        mmm.adstock_decay_spec.organic_media, expected_organic_media
+    )
+    self.assertSequenceEqual(
+        mmm.adstock_decay_spec.organic_rf, expected_organic_rf
+    )
+
+  @parameterized.product(
+      **data_test_utils.ADSTOCK_DECAY_SPEC_CASES,
+      has_undefined_media_channel=(True, False),
+      has_undefined_rf_channel=(True, False),
+      has_undefined_organic_media_channel=(True, False),
+      has_undefined_organic_rf_channel=(True, False),
+      )
+  def test_from_channels_some_undefined(
+      self,
+      media,
+      rf,
+      organic_media,
+      organic_rf,
+      has_undefined_media_channel,
+      has_undefined_rf_channel,
+      has_undefined_organic_media_channel,
+      has_undefined_organic_rf_channel,
+  ):
+    """Test if adstock decay functions are not explicitly passed for all channels."""
+    if not (
+        media or rf or has_undefined_media_channel or has_undefined_rf_channel
+    ):
+      self.skipTest("Invalid test case: Meridian requires paid media.")
+
+    if not sum((
+        has_undefined_media_channel,
+        has_undefined_rf_channel,
+        has_undefined_organic_media_channel,
+        has_undefined_organic_rf_channel
+        )):
+      self.skipTest("Redundant test case: no undefined channels.")
+
+    inp_data = data_test_utils.sample_input_data_revenue(
+        n_media_channels=len(media) + has_undefined_media_channel,
+        n_rf_channels=len(rf) + has_undefined_rf_channel,
+        n_organic_media_channels=len(organic_media)
+        + has_undefined_organic_media_channel,
+        n_organic_rf_channels=len(organic_rf)
+        + has_undefined_organic_rf_channel,
+    )
+
+    decay_spec = media | rf | organic_media | organic_rf
+    model_spec = spec.ModelSpec(adstock_decay_spec=decay_spec)
+    mmm = model.Meridian(input_data=inp_data, model_spec=model_spec)
+
+    if media:
+      expected_media = list(media.values())
+
+      if has_undefined_media_channel:
+        expected_media.append(constants.GEOMETRIC_DECAY)
+    elif has_undefined_media_channel:
+      expected_media = [constants.GEOMETRIC_DECAY]
+    else:
+      expected_media = constants.GEOMETRIC_DECAY
+
+    if rf:
+      expected_rf = list(rf.values())
+
+      if has_undefined_rf_channel:
+        expected_rf.append(constants.GEOMETRIC_DECAY)
+    elif has_undefined_rf_channel:
+      expected_rf = [constants.GEOMETRIC_DECAY]
+    else:
+      expected_rf = constants.GEOMETRIC_DECAY
+
+    if organic_media:
+      expected_organic_media = list(organic_media.values())
+
+      if has_undefined_organic_media_channel:
+        expected_organic_media.append(constants.GEOMETRIC_DECAY)
+    elif has_undefined_organic_media_channel:
+      expected_organic_media = [constants.GEOMETRIC_DECAY]
+    else:
+      expected_organic_media = constants.GEOMETRIC_DECAY
+
+    if organic_rf:
+      expected_organic_rf = list(organic_rf.values())
+
+      if has_undefined_organic_rf_channel:
+        expected_organic_rf.append(constants.GEOMETRIC_DECAY)
+    elif has_undefined_organic_rf_channel:
+      expected_organic_rf = [constants.GEOMETRIC_DECAY]
+    else:
+      expected_organic_rf = constants.GEOMETRIC_DECAY
+
+    self.assertSequenceEqual(mmm.adstock_decay_spec.media, expected_media)
+    self.assertSequenceEqual(mmm.adstock_decay_spec.rf, expected_rf)
+    self.assertSequenceEqual(
+        mmm.adstock_decay_spec.organic_media, expected_organic_media
+    )
+    self.assertSequenceEqual(
+        mmm.adstock_decay_spec.organic_rf, expected_organic_rf
+    )
+
+  @parameterized.product(**data_test_utils.ADSTOCK_DECAY_SPEC_CASES)
+  def test_from_channel_explicit_media_name(
+      self,
+      media,
+      rf,
+      organic_media,
+      organic_rf,
+  ):
+    """Test if one media channel has the name "media"."""
+
+    if not (media or rf):
+      self.skipTest("Invalid test case: Meridian requires paid media.")
+
+    media = media | {"media": constants.BINOMIAL_DECAY}
+
+    inp_data = data_test_utils.sample_input_data_revenue(
+        n_media_channels=len(media),
+        n_rf_channels=len(rf),
+        n_organic_media_channels=len(organic_media),
+        n_organic_rf_channels=len(organic_rf),
+        explicit_media_channel_names=list(media.keys()),
+    )
+
+    decay_spec = media | rf | organic_media | organic_rf
+    model_spec = spec.ModelSpec(adstock_decay_spec=decay_spec)
+    mmm = model.Meridian(input_data=inp_data, model_spec=model_spec)
+
+    expected_media = list(media.values()) or constants.GEOMETRIC_DECAY
+    expected_rf = list(rf.values()) or constants.GEOMETRIC_DECAY
+    expected_organic_media = (
+        list(organic_media.values()) or constants.GEOMETRIC_DECAY
+    )
+    expected_organic_rf = list(organic_rf.values()) or constants.GEOMETRIC_DECAY
+
+    self.assertSequenceEqual(mmm.adstock_decay_spec.media, expected_media)
+    self.assertSequenceEqual(mmm.adstock_decay_spec.rf, expected_rf)
+    self.assertSequenceEqual(
+        mmm.adstock_decay_spec.organic_media, expected_organic_media
+    )
+    self.assertSequenceEqual(
+        mmm.adstock_decay_spec.organic_rf, expected_organic_rf
+    )
+
+  @parameterized.product(
+      **data_test_utils.ADSTOCK_DECAY_SPEC_CASES,
+      bad_channel=({
+          "nonexistent_channel": constants.GEOMETRIC_DECAY
+          },)
+      )
+  def test_from_channels_misnamed_channel_raises_error(
+      self,
+      media,
+      rf,
+      organic_media,
+      organic_rf,
+      bad_channel
+  ):
+    """Test if an exception is raised with an unrecognized channel."""
+    if not (media or rf):
+      self.skipTest("Invalid test case: Meridian requires paid media.")
+
+    inp_data = data_test_utils.sample_input_data_revenue(
+        n_media_channels=len(media),
+        n_rf_channels=len(rf),
+        n_organic_media_channels=len(organic_media),
+        n_organic_rf_channels=len(organic_rf),
+    )
+
+    decay_spec = media | rf | organic_media | organic_rf | bad_channel
+    model_spec = spec.ModelSpec(adstock_decay_spec=decay_spec)
+
+    valid_channel_names = tuple(
+        (media | rf | organic_media | organic_rf).keys()
+    )
+
+    mmm = model.Meridian(input_data=inp_data, model_spec=model_spec)
+
+    with self.assertRaisesWithLiteralMatch(
+        ValueError,
+        "Unrecognized channel names found in `adstock_decay_spec` keys "
+        f"{tuple(decay_spec.keys())}. Keys should either contain only "
+        f"channel_names {valid_channel_names} or be "
+        "one or more of {'media', 'rf', 'organic_media', 'organic_rf'}.",
+    ):
+      _ = mmm.adstock_decay_spec
 
 
 if __name__ == "__main__":