+++ meridian/model/spec.py
@@ -14,6 +14,7 @@
 
 """Defines model specification parameters for Meridian."""
 
+from collections.abc import Mapping
 import dataclasses
 from typing import Sequence
 import warnings
@@ -201,8 +202,15 @@ class ModelSpec:
       `(n_non_media_channels,)` indicating the non-media variables for which the
       non-media value will be scaled by population. If `None`, then no non-media
       variables are scaled by population. Default: `None`.
-    adstock_decay_function: A string to specify the Adstock decay function.
-      Allowed values are 'geometric' and 'binomial'. Default is 'geometric'.
+    adstock_decay_spec: A string or mapping specifying the adstock decay
+      function for each media, RF, organic media and organic RF channel.
+      * If a string, must be either `'geometric'` or `'binomial'`, specifying
+        that decay function for all channels.
+      * If a mapping, keys should be channel names and values should be
+        `'geometric'` or `'binomial'`, with each key-value pair denoting the
+        adstock decay function to use for that channel. Channels that are not
+        specified in the mapping default to using 'geometric'.
+      Default: `'geometric'`.
   """
 
   prior: prior_distribution.PriorDistribution = dataclasses.field(
@@ -228,7 +236,7 @@ class ModelSpec:
   holdout_id: np.ndarray | None = None
   control_population_scaling_id: np.ndarray | None = None
   non_media_population_scaling_id: np.ndarray | None = None
-  adstock_decay_function: str = constants.GEOMETRIC_DECAY
+  adstock_decay_spec: str | Mapping[str, str] = constants.GEOMETRIC_DECAY
 
   def __post_init__(self):
     # Validate media_effects_dist.